Refactor: Extract common vibration code and use it #2264
