+++ AlphaWallet/Browser/Coordinators/DappBrowserCoordinator.swift
@@ -377,35 +377,6 @@ final class DappBrowserCoordinator: NSObject, Coordinator {
     }
 }
 
-enum NotificationFeedbackType {
-    case success
-    case warning
-    case error
-
-    var feedbackType: UINotificationFeedbackGenerator.FeedbackType {
-        switch self {
-        case .success:
-            return .success
-        case .warning:
-            return .warning
-        case .error:
-            return .error
-        }
-    }
-}
-
-extension UINotificationFeedbackGenerator {
-
-    static func show(feedbackType result: NotificationFeedbackType) {
-        let feedbackGenerator = UINotificationFeedbackGenerator()
-        feedbackGenerator.prepare()
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
-            feedbackGenerator.notificationOccurred(result.feedbackType)
-        }
-    }
-}
-
 extension DappBrowserCoordinator: BrowserViewControllerDelegate {
     func didCall(action: DappAction, callbackID: Int, inBrowserViewController viewController: BrowserViewController) {
         guard case .real(let account) = session.account.type else {
