+++ AlphaWallet/UI/WhereIsWalletAddressFoundOverlayView.swift
@@ -66,10 +66,7 @@ class WhereIsWalletAddressFoundOverlayView: UIView {
             self.dialog.transform = .identity
         })
 
-        //TODO sound too
-        let feedbackGenerator = UINotificationFeedbackGenerator()
-        feedbackGenerator.prepare()
-        feedbackGenerator.notificationOccurred(.success)
+        UINotificationFeedbackGenerator.show(feedbackType: .success)
     }
 }
